qa-runner/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
PyYAML==6.0.2
python-multipart==0.0.20
psycopg2-binary==2.9.10
